Move shared settings to it's own file
